added super class Person for Student and Candidate classes
