+++ src/main/java/com/spbgti/dispatcherapp/entity/Candidate.java
@@ -6,21 +6,8 @@ import javax.persistence.*;
 
 
 @Entity
-@Table(name = "SENIOR")
-public class Candidate {
-    @Id
-    @GeneratedValue(generator = "increment")
-    @GenericGenerator(name = "increment", strategy = "increment")
-    private long id;
-
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @Column(name = "surname", nullable = false)
-    private String surname;
-
-    @Column(name = "patronymic")
-    private String patronymic;
+@DiscriminatorValue(value="person")
+public class Candidate extends Person{
 
     @ManyToOne
     @JoinColumn(name = "speciality_id", nullable = false)
@@ -29,42 +16,9 @@ public class Candidate {
     public Candidate() {
     }
 
-    public Candidate(String name, String surname, String patronymic) {
-        this.name = name;
-        this.surname = surname;
-        this.patronymic = patronymic;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getSurname() {
-        return surname;
-    }
-
-    public void setSurname(String surname) {
-        this.surname = surname;
-    }
-
-    public String getPatronymic() {
-        return patronymic;
-    }
-
-    public void setPatronymic(String patronymic) {
-        this.patronymic = patronymic;
+    public Candidate(String name, String surname, String patronymic, Speciality speciality) {
+        super(name, surname, patronymic);
+        this.speciality = speciality;
     }
 
     public Speciality getSpeciality() {
