+++ src/main/java/com/spbgti/dispatcherapp/entity/Student.java
@@ -1,25 +1,10 @@
 package com.spbgti.dispatcherapp.entity;
 
-import org.hibernate.annotations.GenericGenerator;
-
 import javax.persistence.*;
 
 @Entity
-@Table(name = "STUDENT")
-public class Student {
-    @Id
-    @GeneratedValue(generator = "increment")
-    @GenericGenerator(name = "increment", strategy = "increment")
-    private long id;
-
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @Column(name = "surname", nullable = false)
-    private String surname;
-
-    @Column(name = "patronymic")
-    private String patronymic;
+@DiscriminatorValue(value="student")
+public class Student extends Person{
 
     @ManyToOne
     @JoinColumn(name = "group_id", nullable = false)
@@ -29,42 +14,8 @@ public class Student {
     }
 
     public Student(String name, String surname, String patronymic, Group group) {
-        this.name = name;
-        this.surname = surname;
-        this.patronymic = patronymic;
+        super(name, surname, patronymic);
         this.group = group;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getSurname() {
-        return surname;
-    }
-
-    public void setSurname(String surname) {
-        this.surname = surname;
-    }
-
-    public String getPatronymic() {
-        return patronymic;
-    }
-
-    public void setPatronymic(String patronymic) {
-        this.patronymic = patronymic;
     }
 
     public Group getGroup() {
